Quick fix of test structure and README update
